+++ ACG/ACG/ObjLastWall.cpp
@@ -12,8 +12,8 @@ using namespace GameL;
 //コンストラクタ
 CObjLastWall::CObjLastWall(int x, int y)
 {
-	m_px = x * 64;
-	m_py = y * 64;
+	m_px = (float)x * 64.0f;
+	m_py = (float)y * 64.0f;
 }
 
 //イニシャライズ
@@ -21,7 +21,7 @@ void CObjLastWall::Init()
 {
 	m_ani_time = 0;
 	m_ani_frame = 1;  //静止フレームを初期にする
-	m_ani_max_time = 17; //アニメーション間隔幅
+	m_ani_max_time = 17.0f; //アニメーション間隔幅
 	m_ani_start = false;
 
 	//当たり判定																
@@ -51,7 +51,7 @@ void CObjLastWall::Action()
 	//}
 
 	//HitBoxの位置を更新する
-	HitBoxUpData(Hits::GetHitBox(this), m_px+30, m_py+255);
+	HitBoxUpData(Hits::GetHitBox(this), m_px+30.0f, m_py+255.0f);
 }
 
 //ドロー
@@ -72,10 +72,10 @@ void CObjLastWall::Draw()
 	src.m_bottom = 64.0f;
 
 	//描画位置
-	dst.m_top = m_py - objmap->GetScrollY()+21;
+	dst.m_top = m_py - objmap->GetScrollY()+21.0f;
 	dst.m_left = m_px - objmap->GetScrollX();
-	dst.m_right = dst.m_left + 96;
-	dst.m_bottom = dst.m_top + 256;
+	dst.m_right = dst.m_left + 96.0f;
+	dst.m_bottom = dst.m_top + 256.0f;
 
 	//描画(上の部分)
 	Draw::Draw(22, &src, &dst, color, 0.0f);
@@ -87,10 +87,10 @@ void CObjLastWall::Draw()
 	src.m_bottom = 64.0f;
 
 	//描画位置
-	dst.m_top = m_py - objmap->GetScrollY()+235;
-	dst.m_left = m_px - objmap->GetScrollX()+30;
-	dst.m_right = dst.m_left + 32;
-	dst.m_bottom = dst.m_top + 512;
+	dst.m_top = m_py - objmap->GetScrollY()+235.0f;
+	dst.m_left = m_px - objmap->GetScrollX()+30.0f;
+	dst.m_right = dst.m_left + 32.0f;
+	dst.m_bottom = dst.m_top + 512.0f;
 	//描画(下の部分)
 	Draw::Draw(21, &src, &dst, color, 0.0f);
 }
